+++ tests/test_commands/test_pack_command.py
@@ -1,6 +1,5 @@
 from mock import patch
 from pyfakefs import fake_filesystem_unittest
-from shellfoundry.utilities.python_depedencies_packager import PythonDependenciesPackager
 from tests.asserts import *
 from shellfoundry.commands.pack_command import PackCommandExecutor
 
@@ -10,8 +9,8 @@ class TestPackCommandExecutor(fake_filesystem_unittest.TestCase):
         self.setUpPyfakefs()
 
     @patch('click.echo')
-    @patch('shellfoundry.utilities.python_depedencies_packager.pip')
-    def test_build_package_package_created(self, pip_mock,echo_mock):
+    @patch('shellfoundry.utilities.python_dependencies_packager.pip')
+    def test_build_package_package_created(self, pip_mock, echo_mock):
         # Arrange
         self.fs.CreateFile('nut_shell/shell.yml', contents="""
 shell:
@@ -37,8 +36,8 @@ shell:
         echo_mock.assert_any_call(u'Shell package was successfully created:')
 
     @patch('click.echo')
-    @patch('shellfoundry.utilities.python_depedencies_packager.pip')
-    def test_proper_error_message_displayed_when_shell_yml_is_in_wrong_format(self,pip_mock, echo_mock):
+    @patch('shellfoundry.utilities.python_dependencies_packager.pip')
+    def test_proper_error_message_displayed_when_shell_yml_is_in_wrong_format(self, pip_mock, echo_mock):
         # Arrange
         self.fs.CreateFile('nut_shell/shell.yml', contents='WRONG YAML FORMAT')
         os.chdir('nut_shell')
@@ -52,8 +51,8 @@ shell:
         echo_mock.assert_any_call(u'shell.yml format is wrong')
 
     @patch('click.echo')
-    @patch('shellfoundry.utilities.python_depedencies_packager.pip')
-    def test_proper_error_message_displayed_when_shell_yml_missing(self,pip_mock, echo_mock):
+    @patch('shellfoundry.utilities.python_dependencies_packager.pip')
+    def test_proper_error_message_displayed_when_shell_yml_missing(self, pip_mock, echo_mock):
         # Arrange
         self.fs.CreateFile('nut_shell/datamodel/datamodel.xml')
         os.chdir('nut_shell')
@@ -65,3 +64,25 @@ shell:
 
         # Assert
         echo_mock.assert_any_call(u'shell.yml file is missing')
+
+    @patch('shellfoundry.commands.pack_command.ShellPackageBuilder.pack')
+    def test_tosca_based_shell_packager_called_when_shell_contains_tosca_meta_file(self, pack_mock):
+        # Arrange
+        self.fs.CreateFile('nut-shell/TOSCA-Metadata/TOSCA.meta',
+                           contents='TOSCA-Meta-File-Version: 1.0 \n'
+                                    'CSAR-Version: 1.1 \n'
+                                    'Created-By: Anonymous \n'
+                                    'Entry-Definitions: shell-definition.yml')
+
+        self.fs.CreateFile('nut-shell/shell-definition.yml',
+                           contents='SOME SHELL DEFINITION')
+
+        os.chdir('nut-shell')
+
+        command_executor = PackCommandExecutor()
+
+        # Act
+        command_executor.pack()
+
+        # Assert
+        self.assertTrue(pack_mock.called)